Implementing a geoblacklight:webpacker generator in order to facilitate with the upgrade process; Removing the overridden Blacklight base template; Updating the README
