+++ template.rb
@@ -1,17 +1,11 @@
-gem 'blacklight', '>= 6.11'
-gem 'geoblacklight', '>= 1.9'
+gem 'blacklight', '~> 7.0'
+gem 'geoblacklight', git: 'https://github.com/geoblacklight/geoblacklight.git'
 gem 'webpacker', '~> 3.5'
 
 run 'bundle install'
 
 generate 'blacklight:install', '--devise'
 generate 'geoblacklight:install', '-f'
+generate 'geoblacklight:webpacker', '-f'
 
-run 'bundle exec rails webpacker:install'
-run 'yarn --version'
-if $CHILD_STATUS.exitstatus.zero?
-  run 'bundle exec rails yarn:install'
-else
-  run 'npm install'
-end
 rake 'db:migrate'
